+++ rust/operator-binary/src/backend/dynamic.rs
@@ -7,7 +7,7 @@ use std::{collections::HashSet, fmt::Display};
 
 use super::{
     kerberos_keytab::{self, KerberosProfile},
-    pod_info::PodInfo,
+    pod_info::{PodInfo, SchedulingPodInfo},
     tls, SecretBackend, SecretBackendError, SecretVolumeSelector,
 };
 use crate::crd::{self, SecretClass};
@@ -51,9 +51,10 @@ impl<B: SecretBackend + Send + Sync> SecretBackend for DynamicAdapter<B> {
     async fn get_qualified_node_names(
         &self,
         selector: &SecretVolumeSelector,
+        pod_info: SchedulingPodInfo,
     ) -> Result<Option<HashSet<String>>, Self::Error> {
         self.0
-            .get_qualified_node_names(selector)
+            .get_qualified_node_names(selector, pod_info)
             .await
             .map_err(|err| DynError(Box::new(err)))
     }